massive update
updates to collision, speed up reader for mod loading, also fix AdjustChannel/SetChannelAttributes

+++ RSDKv5/RSDK/Audio/Audio.cpp
@@ -514,7 +514,7 @@ int32 RSDK::PlaySfx(uint16 sfx, uint32 loopPoint, uint32 priority)
     return slot;
 }
 
-void RSDK::SetChannelAttributes(uint8 channel, float volume, float panning, float speed)
+void RSDK::SetChannelAttributes(uint32 channel, float volume, float panning, float speed)
 {
     OnChannelAttributesChanged(&channel, &volume, &panning, &speed);
     if (channel < CHANNEL_COUNT) {
